fix reference for feature.trash project and reformat fixes: #1778
