+++ starsky/starsky.feature.metaupdate/readme.md
@@ -1,6 +1,7 @@
 # MetaUpdate
 
-The scope of this feature assembly is to provide a single endpoint to get batch of single update the meta data of a single or group of images
-This has an dependency on storage, readmeta, writemeta and database
+The scope of this feature assembly is to provide a single endpoint to get batch of single update the
+meta-data of a single or group of images
+This has a dependency on storage, readmeta, writemeta and database
 
 So when there is an update it should update the source image and the content in the database
