+++ starsky/starsky.foundation.metaupdate/Helpers/AddParentCacheIfNotExist.cs
@@ -1,23 +1,11 @@
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using starsky.foundation.database.Interfaces;
 using starsky.foundation.platform.Helpers;
 using starsky.foundation.platform.Interfaces;
 
-namespace starsky.feature.metaupdate.Helpers
+namespace starsky.foundation.metaupdate.Helpers
 {
-	public class AddParentCacheIfNotExist
+	public class AddParentCacheIfNotExist(IQuery query, IWebLogger logger)
 	{
-		private readonly IQuery _query;
-		private readonly IWebLogger _logger;
-
-		public AddParentCacheIfNotExist(IQuery query, IWebLogger logger)
-		{
-			_query = query;
-			_logger = logger;
-		}
-
 		internal async Task<List<string>> AddParentCacheIfNotExistAsync(IEnumerable<string> updatedPaths)
 		{
 			var parentDirectoryList = new HashSet<string>();
@@ -28,22 +16,22 @@ namespace starsky.feature.metaupdate.Helpers
 			}
 
 			var shouldAddParentDirectoriesToCache = parentDirectoryList.Where(parentDirectory =>
-				!_query.CacheGetParentFolder(parentDirectory).Item1).ToList();
+				!query.CacheGetParentFolder(parentDirectory).Item1).ToList();
 
 			if ( shouldAddParentDirectoriesToCache.Count == 0 )
 			{
 				return new List<string>();
 			}
 
-			var databaseQueryResult = await _query.GetAllObjectsAsync(shouldAddParentDirectoriesToCache);
+			var databaseQueryResult = await query.GetAllObjectsAsync(shouldAddParentDirectoriesToCache);
 
-			_logger.LogInformation("[AddParentCacheIfNotExist] files added to cache " +
+			logger.LogInformation("[AddParentCacheIfNotExist] files added to cache " +
 								   string.Join(",", shouldAddParentDirectoriesToCache));
 
 			foreach ( var directory in shouldAddParentDirectoriesToCache )
 			{
 				var byDirectory = databaseQueryResult.Where(p => p.ParentDirectory == directory).ToList();
-				_query.AddCacheParentItem(directory, byDirectory);
+				query.AddCacheParentItem(directory, byDirectory);
 			}
 			return shouldAddParentDirectoriesToCache;
 		}
