+++ starsky/starsky.foundation.metaupdate/Interfaces/IMetaReplaceService.cs
@@ -1,8 +1,6 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using starsky.foundation.database.Models;
 
-namespace starsky.feature.metaupdate.Interfaces;
+namespace starsky.foundation.metaupdate.Interfaces;
 
 public interface IMetaReplaceService
 {
