+++ starsky/starsky.foundation.metaupdate/Services/MetaPreflight.cs
@@ -1,12 +1,9 @@
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using starsky.feature.metaupdate.Helpers;
-using starsky.feature.metaupdate.Interfaces;
 using starsky.foundation.database.Helpers;
 using starsky.foundation.database.Interfaces;
 using starsky.foundation.database.Models;
 using starsky.foundation.injection;
+using starsky.foundation.metaupdate.Helpers;
+using starsky.foundation.metaupdate.Interfaces;
 using starsky.foundation.platform.Helpers;
 using starsky.foundation.platform.Interfaces;
 using starsky.foundation.platform.Models;
@@ -14,7 +11,7 @@ using starsky.foundation.storage.Interfaces;
 using starsky.foundation.storage.Models;
 using starsky.foundation.storage.Storage;
 
-namespace starsky.feature.metaupdate.Services
+namespace starsky.foundation.metaupdate.Services
 {
 	[Service(typeof(IMetaPreflight), InjectionLifetime = InjectionLifetime.Scoped)]
 	public class MetaPreflight : IMetaPreflight
