+++ starsky/starsky.foundation.metaupdate/Services/MetaUpdateService.cs
@@ -1,14 +1,9 @@
-using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.IO;
-using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Threading.Tasks;
-using starsky.feature.metaupdate.Interfaces;
 using starsky.foundation.database.Interfaces;
 using starsky.foundation.database.Models;
 using starsky.foundation.injection;
+using starsky.foundation.metaupdate.Interfaces;
 using starsky.foundation.platform.Helpers;
 using starsky.foundation.platform.Interfaces;
 using starsky.foundation.platform.Models;
@@ -23,7 +18,7 @@ using ExifToolCmdHelper = starsky.foundation.writemeta.Helpers.ExifToolCmdHelper
 
 [assembly: InternalsVisibleTo("starskytest")]
 
-namespace starsky.feature.metaupdate.Services;
+namespace starsky.foundation.metaupdate.Services;
 
 [Service(typeof(IMetaUpdateService), InjectionLifetime = InjectionLifetime.Scoped)]
 public class MetaUpdateService : IMetaUpdateService
