+++ starsky/starsky.sln
@@ -176,6 +176,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "starskydependenciesdownload
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "starsky.foundation.geo", "starsky.foundation.geo\starsky.foundation.geo.csproj", "{D6D53732-2CEC-489D-BE54-E15AEE30B6D5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "starsky.foundation.metaupdate", "starsky.foundation.metaupdate\starsky.foundation.metaupdate.csproj", "{9B6E6A97-9F50-47A2-995B-B54895880573}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -386,6 +388,10 @@ Global
 		{D6D53732-2CEC-489D-BE54-E15AEE30B6D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D6D53732-2CEC-489D-BE54-E15AEE30B6D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D6D53732-2CEC-489D-BE54-E15AEE30B6D5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9B6E6A97-9F50-47A2-995B-B54895880573}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9B6E6A97-9F50-47A2-995B-B54895880573}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9B6E6A97-9F50-47A2-995B-B54895880573}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9B6E6A97-9F50-47A2-995B-B54895880573}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -449,5 +455,6 @@ Global
 		{C6F6B684-F8EB-4DFD-9143-33F0262153FE} = {1C1EB4A5-08D0-4014-AE1F-962642A4E5D3}
 		{A5BE4F80-975C-4F15-AB99-D046E6F8726D} = {B974FC20-C3EE-4EB0-AF25-F0D8DA2C28D7}
 		{D6D53732-2CEC-489D-BE54-E15AEE30B6D5} = {1C1EB4A5-08D0-4014-AE1F-962642A4E5D3}
+		{9B6E6A97-9F50-47A2-995B-B54895880573} = {1C1EB4A5-08D0-4014-AE1F-962642A4E5D3}
 	EndGlobalSection
 EndGlobal
