+++ starsky/starskytest/Controllers/MetaUpdateControllerTest.cs
@@ -13,12 +13,12 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using starsky.Controllers;
-using starsky.feature.metaupdate.Interfaces;
-using starsky.feature.metaupdate.Services;
 using starsky.feature.realtime.Interface;
 using starsky.foundation.database.Data;
 using starsky.foundation.database.Models;
 using starsky.foundation.database.Query;
+using starsky.foundation.metaupdate.Interfaces;
+using starsky.foundation.metaupdate.Services;
 using starsky.foundation.platform.Extensions;
 using starsky.foundation.platform.Helpers;
 using starsky.foundation.platform.Interfaces;
