+++ starsky/starskytest/starsky.feature.trash/Services/MoveToTrashServiceTest.cs
@@ -6,12 +6,12 @@ using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using starsky.feature.metaupdate.Services;
 using starsky.feature.trash.Services;
 using starsky.foundation.database.Data;
 using starsky.foundation.database.Models;
 using starsky.foundation.database.Query;
 using starsky.foundation.database.Thumbnails;
+using starsky.foundation.metaupdate.Services;
 using starsky.foundation.native.Trash;
 using starsky.foundation.platform.Models;
 using starsky.foundation.readmeta.Services;
