+++ starsky/starskytest/starsky.foundation.accountmanagement/Services/UserManagerTest.cs
@@ -9,11 +9,11 @@ using Microsoft.EntityFrameworkCore.Storage;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using starsky.feature.metaupdate.Services;
 using starsky.foundation.accountmanagement.Interfaces;
 using starsky.foundation.accountmanagement.Services;
 using starsky.foundation.database.Data;
 using starsky.foundation.database.Models.Account;
+using starsky.foundation.metaupdate.Services;
 using starsky.foundation.platform.Models;
 using starskytest.FakeMocks;
