+++ starsky/starskytest/starsky.foundation.metaupdate/Services/MetaPreflightTest.cs
@@ -3,13 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using starsky.feature.metaupdate.Services;
 using starsky.foundation.database.Models;
+using starsky.foundation.metaupdate.Services;
 using starsky.foundation.platform.Models;
 using starskytest.FakeCreateAn;
 using starskytest.FakeMocks;
 
-namespace starskytest.starsky.feature.metaupdate.Services;
+namespace starskytest.starsky.foundation.metaupdate.Services;
 
 [TestClass]
 public sealed class MetaPreflightTest
