+++ starsky/starskytest/starsky.foundation.metaupdate/Services/MetaUpdateServiceTest.cs
@@ -7,10 +7,10 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using starsky.feature.metaupdate.Services;
 using starsky.foundation.database.Data;
 using starsky.foundation.database.Models;
 using starsky.foundation.database.Query;
+using starsky.foundation.metaupdate.Services;
 using starsky.foundation.platform.Models;
 using starsky.foundation.platform.Thumbnails;
 using starsky.foundation.readmeta.Services;
@@ -18,7 +18,7 @@ using starsky.foundation.storage.Storage;
 using starskytest.FakeCreateAn;
 using starskytest.FakeMocks;
 
-namespace starskytest.starsky.feature.metaupdate.Services;
+namespace starskytest.starsky.foundation.metaupdate.Services;
 
 [TestClass]
 public sealed class MetaUpdateServiceTest
